geo-engine/requirements_db.txt
# Tamil Nadu Dataset & Database Dependencies
# Install with:
#   pip install -r requirements_db.txt

# Geospatial
osmnx>=1.7.1
geopandas>=0.14.1
shapely>=2.0.2
pyproj>=3.6.0
fiona>=1.9.5

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# Data processing
pandas>=2.1.3
numpy>=1.26.0
networkx>=3.2.1

# HTTP / API calls
requests>=2.31.0

# Environment config
python-dotenv>=1.0.0

# Progress display
tqdm>=4.66.1
